+++ stark/extract.py
@@ -73,7 +73,7 @@ def flux_coo(frame, variance, ord_pos, ap_rad):
 
     Parameters
     ----------
-    frame: ndarray
+    frame : ndarray
         3D array containing data, [nints, nrows, ncols]
     variance : ndarray
         3D array as same size as `frame` containing variance of the data.
@@ -174,6 +174,121 @@ def norm_flux_coo(pix_array, col_array_pos, spec = None):
             norm_array[ind0:ind1,1] = pix_array[ind0:ind1,1]/norm_sum
             norm_array[ind0:ind1,2] = pix_array[ind0:ind1,2]/norm_sum**2
     return norm_array
+
+def univariate_psf_frame(data, ord_pos, ap_rad, pix_array, **kwargs):
+    """To generate PSF frame for the data
+    
+    Given frame, order position, aperture radius and normalized pixel array,
+    this function derives a single PSF for whole data cube and use it to 
+    generate pixel-sampled PSFs for each column.
+
+    Parameters
+    ----------
+    data : ndarray
+        3D array containing data, [nints, nrows, ncols]
+    order_pos : ndarray
+        2D array, with shape [nints, ncols] contaning pixel positions of order
+    ap_rad : float
+        Radius of the aperture to consider
+    pix_array : ndarray
+        Normalized pixel positions, as computed from `norm_flux_coo`.
+    **kwargs :
+        Additional keywords provided to `fit_spline_univariate` function and to LSQUnivariateSpline
+    
+    Returns
+    -------
+    frame : ndarray
+        Data cube containing pixel-sampled PSF for each column
+    """
+    frame = np.copy(data)
+    
+    nints = frame.shape[0]
+    ncols = frame.shape[2]
+
+    # To sort array
+    sortarg =  np.argsort(pix_array[:,0])
+    sort_arr = pix_array[sortarg,:]
+    psf_spline, mask = fit_spline_univariate(sort_arr, **kwargs)
+
+    for integration in range(nints):
+        for col in range(ncols):
+            if ord_pos[integration, col] < 0 or ord_pos[integration, col] >= frame.shape[1]:
+                continue
+            i0 =  int(round(ord_pos[integration, col] - ap_rad))
+            i1 = int(round(ord_pos[integration, col] + ap_rad))
+            if i0 < 0:
+                i0 = 0
+            if i1 >= frame.shape[1]:
+                i1 = frame.shape[1] - 1
+
+            x = np.arange(i0,i1) - ord_pos[integration, col]
+            frame[integration, i0:i1, col] = np.maximum(psf_spline(x), 0) # Enforce positivity
+    #        frame[integration, i0:i1, col] /= np.sum(frame[integration, i0:i1, col]) # Enforce normalisation (why commented)
+    return frame
+
+def psf_extract(psf_frame, data, variance, mask, ord_pos, ap_rad):
+    """Use derived PSF frame (the psf sampled on the image) to extract the spectrum.
+
+    This function fits the derived PSF frame to the actual data to extract the
+    spectrum and variance on it.
+
+    Parameters
+    ----------
+    psf_frame : ndarray
+        PSF frame with dimension [nints, nrows, ncols]
+    data : ndarray
+        3D array containing data, [nints, nrows, ncols]
+    variance : ndarray
+        Variance on the data frame, same shape as the `data` array
+    mask : ndarray
+        Array containing mask; only those points with value = True will be considered in extraction
+    ord_pos : ndarray
+        2D array, with shape [nints, ncols] contaning pixel positions of order
+    ap_rad : float
+        Radius of the aperture to consider
+    
+    Returns
+    -------
+    spec : ndarray
+        Extracted flux as a matrix with format [nints, ncols]
+    var : ndarray
+        Variance of each fluc point in the same format as `spec`.
+    synth : ndarray
+        Synthetic image constructed using the flux and the PSF frame;
+        useful for producing a residual image.
+    """
+
+    ncols = data.shape[2]
+    nints = data.shape[0]
+
+    spec = np.zeros((nints, ncols))
+    var = np.zeros((nints, ncols))
+    synth = np.zeros(data.shape)
+
+    for integration in range(nints):
+        for col in range(ncols):
+            if ord_pos[integration, col] < 0 or ord_pos[integration, col] >= data.shape[1]:
+                continue
+            i0 =  int(round(ord_pos[integration, col] - ap_rad))
+            i1 = int(round(ord_pos[integration, col] + ap_rad))
+
+            if i0 < 0:
+                i0 = 0
+            if i1 >= data.shape[1]:
+                i1 = data.shape[1] - 1
+ 
+            ind = np.array(range(i0,i1))
+            mask2 = psf_frame[integration, ind, col] > 0
+            ind = ind[mask2]
+            denom = np.sum(mask[ind, col] * psf_frame[integration, ind, col]**2 / 
+                           variance[integration, ind, col])
+            with warnings.catch_warnings():
+                warnings.simplefilter("ignore")
+                spec[integration, col] = (np.sum(mask[ind, col] * psf_frame[integration, ind, col] *
+                    data[integration, ind, col] / variance[integration, ind, col]) / denom )
+                var[integration, col] = np.sum(mask[ind, col] * psf_frame[integration, ind, col]) / denom
+            synth[integration, i0:i1, col] = spec[integration, col] * psf_frame[integration, i0:i1, col]
+    return spec, var, synth
 
 def fit_spline_univariate(pixel_sorted, oversample=1, clip=5, niters=3, **kwargs):
     """Fit a Univariate spline to pixel arrays